make it pedantic

+++ csapp.c
@@ -569,7 +569,7 @@ ssize_t rio_readnb(rio_t *rp, void *usrbuf, size_t n) {
  */
 /* $begin rio_readlineb */
 ssize_t rio_readlineb(rio_t *rp, void *usrbuf, size_t maxlen) {
-    int n, rc;
+    size_t n, rc;
     char c, *bufp = usrbuf;
 
     for (n = 1; n < maxlen; n++) {
